requirements.txt
# Requirements for WhisperX GPU Diarization
# ==========================================

# Core dependencies - compatible versions
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.2

# WhisperX, Speechbrain and related
whisperx>=3.1.1
faster-whisper>=0.10.0

# Environment variables
python-dotenv>=1.0.0

# Diarization
pyannote.audio>=3.1.1
speechbrain>=0.5.16

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0

# Utilities
numpy<2.0.0
pandas>=2.0.0
tqdm>=4.65.0

transformers 
accelerate 
bitsandbytes

#STT Detection
webrtcvad
wavio
sounddevice